Consolidate parsing runtime fields from source
Every runtime field type allows users to omit its script, in which case the field values will be loaded at runtime from _source.

This is implemented by having each field type expose a parse from source script factory that extracts the values and converts them to the appropriate type that can then be emitted.

The extraction logic from source is though always the same, what changes between the different types is the factory type that is needed and how the object values are converted to their appropriate type. This commit moves the common bits to AbstractFieldScript. Especially the conversion from Object to the appropriate type is handy in a specific method as it will be reused to emit multiple fields from a single
script.
