+++ server/src/main/java/org/elasticsearch/script/LongFieldScript.java
@@ -9,6 +9,7 @@
 package org.elasticsearch.script;
 
 import org.apache.lucene.index.LeafReaderContext;
+import org.elasticsearch.index.mapper.NumberFieldMapper;
 import org.elasticsearch.search.lookup.SearchLookup;
 
 import java.util.Map;
@@ -29,6 +30,15 @@ public abstract class LongFieldScript extends AbstractLongFieldScript {
 
     public LongFieldScript(String fieldName, Map<String, Object> params, SearchLookup searchLookup, LeafReaderContext ctx) {
         super(fieldName, params, searchLookup, ctx);
+    }
+
+    @Override
+    protected void emitFromObject(Object v) {
+        try {
+            emit(NumberFieldMapper.NumberType.objectToLong(v, true));
+        } catch (Exception e) {
+            // ignore;
+        }
     }
 
     public static class Emit {
